requirements.txt
rich
pymongo
grpcio
grpcio-tools
ccdexplorer-fundamentals
betterproto==2.0.0b5
python-dotenv